requirements.txt
bcrypt==5.0.0
blinker==1.9.0
click==8.3.1
colorama==0.4.6
Flask==3.1.3
Flask-Bcrypt==1.0.1
greenlet==3.3.2
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
psycopg==3.3.3
psycopg-binary==3.3.3
psycopg2-binary==2.9.11
PyJWT==2.12.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
tzdata==2025.3
Werkzeug==3.1.6
